Refactor RGB distribution calculation
